tools/bgcinfra/requirements.txt
Jinja2==2.10.1
MarkupSafe==1.1.1
PyYAML==5.1.1
